backend/ai_engine/requirements.txt
fastapi
uvicorn
python-dotenv
groq
pytesseract
opencv-python
python-docx
pdfplumber
speechrecognition
pydantic
python-multipart